CultivateBackend/requirements.txt
flask
google-genai
serpapi
google-search-results
json